wip: working post and get routes

+++ src/database/migrations/1587228331051-CreateTransactions.ts
@@ -20,7 +20,7 @@ export default class CreateTransactions1587228331051
           },
           {
             name: 'value',
-            type: 'money',
+            type: 'float',
           },
           {
             name: 'type',
